Updated defaults and Program.cs to work with whitespace lexing
Fixed bug with LookbackEnumerator that caused crash on ForgetPrevious
when there were no characters to forget

DelimiterLexer no longer recognizes whitespace delimiters by default

Program.cs evaluates a string with double whitespace

+++ MathParser/Util/LookbackEnumerator.cs
@@ -58,6 +58,9 @@ namespace MathParser.Util
 
         public void ForgetPreceding()
         {
+            if (CurrentIndex < 0)
+                return;
+
             LoadedValues.RemoveRange(0, CurrentIndex);
             CurrentIndex = -1;
         }
